+++ libraries/ts-command-line/src/providers/CommandLineParser.ts
@@ -6,6 +6,7 @@ import * as colors from 'colors';
 
 import { CommandLineAction } from './CommandLineAction';
 import { CommandLineParameterProvider, ICommandLineParserData } from './CommandLineParameterProvider';
+import { CommandLineParserExitError, CustomArgumentParser} from './CommandLineParserExitError';
 
 /**
  * Options for the {@link CommandLineParser} constructor.
@@ -21,37 +22,6 @@ export interface ICommandLineParserOptions {
    * General documentation that is included in the "--help" main page
    */
   toolDescription: string;
-}
-
-export class CommandLineParserExitError extends Error {
-  public readonly exitCode: number;
-
-  public constructor(exitCode: number, message: string) {
-    super(message);
-
-    // Manually set the prototype, as we can no longer extend built-in classes like Error, Array, Map, etc
-    // https://github.com/microsoft/TypeScript-wiki/blob/master/Breaking-Changes.md#extending-built-ins-like-error-array-and-map-may-no-longer-work
-    //
-    // Note: the prototype must also be set on any classes which extend this one
-    (this as any).__proto__ = CommandLineParserExitError.prototype; // eslint-disable-line @typescript-eslint/no-explicit-any
-
-    this.exitCode = exitCode;
-  }
-}
-
-class CustomArgumentParser extends argparse.ArgumentParser {
-  public exit(status: number, message: string): void { // override
-    throw new CommandLineParserExitError(status, message);
-  }
-
-  public error(err: Error | string): void { // override
-    // Ensure the ParserExitError bubbles up to the top without any special processing
-    if (err instanceof CommandLineParserExitError) {
-      throw err;
-    }
-
-    super.error(err);
-  }
 }
 
 /**
